% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{aggSimilarity}
\alias{aggSimilarity}
\title{Aggregate similarity}
\usage{
aggSimilarity(porftolios, simM = NULL, weights = NULL, rw = NULL,
  arc = FALSE)
}
\arguments{
\item{simM}{an optional (n x n) similarity matrix}

\item{weights}{a (n x n) matrix of weights for weghting matrix to an average}

\item{rw}{(1 x k) vector of weights on sectors}

\item{arc}{logical; should the cosine similarities be transformed to radians (by arccos) first?}

\item{portfolios}{a (n x k) matrix of portfolios where rows represent n subjects and columns their portfolios}
}
\value{
A numeric scalar value of (weighted) mean of the upper traingle of similarity matrix is returned.
}
\description{
\code{aggSimilarity} computes aggregate similarity
}
\details{
Using \code{arc=TRUE} returns (weighted) mean of 1-(2*arccos(cos(similarity))/pi)
}
\seealso{
\code{\link{aggSimilarity_time}}
}
