% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{aggSimilarity_time}
\alias{aggSimilarity_time}
\title{Aggregate similarity applied to a list}
\usage{
aggSimilarity_time(portfolios_list, ...)
}
\arguments{
\item{portfolios_list}{a list of portfolio matrices where rows represent n subjects and columns their portfolios}

\item{...}{other archments past to \code{\link{aggSimilarity}}}
}
\value{
A numeric vector of the same lenght as input.
}
\description{
\code{aggSimilarity_time} computes aggregate similarities for a given list of portfolios
}
\seealso{
\code{\link{aggSimilarity}}
}
