% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/allPairs.R
\name{allPairsRcpp}
\alias{allPairsRcpp}
\alias{allPairs}
\title{Generate all possible pairs from a vector or a sequence of integer numbers}
\usage{
allPairsRcpp(n)

allPairs(x, unique = TRUE)
}
\arguments{
\item{n}{maximum number of the sequence}
}
\value{
\code{allPairsRcpp} returns (n*(n-1)/2) x 2 integer matrix all unique combinations (pairs) of numbers from a sequence of integers 1..n

\code{allPairs} returns (n*(n-1)/2) x 2 matrix all combinations (pairs) of values from vector \code{x}
}
\description{
Generate all possible pairs from a vector or a sequence of integer numbers
}
