% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{arcCosDist}
\alias{arcCosDist}
\title{arccos distance transformed to [0;1] scale}
\usage{
arcCosDist(x)
}
\arguments{
\item{x}{cosine similarities of positive vectors (as a vector or matrix)}
}
\value{
angular distance
}
\description{
arccos distance transformed to [0;1] scale
}
