% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{cosine}
\alias{cosine}
\title{generic cosine similarity function}
\usage{
cosine(x, y = NULL, weights = NULL, complete = FALSE)
}
\arguments{
\item{x}{numeric (vector or matrix)}

\item{y}{if x is numeric vecor, y has to be numeric vector of the same length}

\item{weights}{numeric; weights for weighted or soft cosine similarity}

\item{complete}{logical; if FALSE (default) only upper triangle of the similarity matrix is computed}
}
\description{
generic cosine similarity function
}
