% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cosSimilarity}
\alias{cosSimilarity}
\title{Compute cosine similarity of two numeric vectors}
\usage{
cosSimilarity(a, b)
}
\arguments{
\item{a}{numeric vector}

\item{b}{numeric vector}
}
\value{
Numeric value of cosine similarity
}
\description{
Compute cosine similarity of two numeric vectors
}

