% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mCosSimilarity}
\alias{mCosSimilarity}
\title{Compute cosine similarity for every pair of rows from given matrix}
\usage{
mCosSimilarity(input_mat)
}
\arguments{
\item{input_mat}{numeric input matrix}
}
\value{
Upper triangle matrix where \{i, j\} element is the cosine similarity of i-th and j-th row of the original matrix.
}
\description{
Compute cosine similarity for every pair of rows from given matrix
}

