% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{randMatrixGamma}
\alias{randMatrixGamma}
\title{Generate random variation of a matrix keeping rows and cols sums}
\usage{
randMatrixGamma(original_mat, max_iter, shape, rate = 1)
}
\arguments{
\item{original_mat}{the original matrix that will be replicated}

\item{max_iter}{maximum number of iterations when coverging to the rows and cols sums}

\item{shape}{parameter of the gamma distribution}

\item{rate}{parameter of the gamma distribution}
}
\value{
numeric matrix with the same dimensions and (almost) rows' and columns' sums as the input matrix
}
\description{
Generate random variation of a matrix keeping rows and cols sums
}

