% Generated by roxygen2 (4.0.1): do not edit by hand
\name{randMatrixUnif}
\alias{randMatrixUnif}
\title{Generate random variation of a matrix keeping rows and cols sums}
\usage{
randMatrixUnif(original_mat, max_iter)
}
\arguments{
\item{original_mat}{the original matrix that will be replicated}

\item{max_iter}{maximum number of iterations when coverging to the rows and cols sums}
}
\value{
numeric matrix with the same dimensions and (almost) rows' and columns' sums as the input matrix
}
\description{
Generate random variation of a matrix keeping rows and cols sums
}

