% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{restRandMat}
\alias{restRandMat}
\title{Generate random variation of a matrix keeping rows and cols sums}
\usage{
restRandMat(original_mat, max_iter, type, par1 = 1, par2 = 1)
}
\arguments{
\item{original_mat}{the original matrix that will be replicated}

\item{max_iter}{maximum number of iterations when coverging to the rows and cols sums}
}
\value{
numeric matrix with the same dimensions and (almost) rows' and columns' sums as the input matrix
}
\description{
Generate random variation of a matrix keeping rows and cols sums
}

