% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{restRandMatLNorm}
\alias{restRandMatLNorm}
\title{Generate random variation of a matrix keeping rows and cols sums. Deprecated, use restRandMat().}
\usage{
restRandMatLNorm(original_mat, max_iter, meanl = 0, sdl = 1)
}
\arguments{
\item{original_mat}{the original matrix that will be replicated}

\item{max_iter}{maximum number of iterations when coverging to the rows and cols sums}

\item{meanl}{parameter of the log-normal distribution}

\item{sdl}{parameter of the log-normal distribution}
}
\value{
numeric matrix with the same dimensions and (almost) rows' and columns' sums as the input matrix
}
\description{
Generate random variation of a matrix keeping rows and cols sums. Deprecated, use restRandMat().
}

