% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sftVCosSimilarity}
\alias{sftVCosSimilarity}
\title{Compute soft weighted cosine similarity of two numeric vectors}
\usage{
sftVCosSimilarity(a, b, weights)
}
\arguments{
\item{a}{numeric vector}

\item{b}{numeric vector}

\item{w}{numeric matrix of weights}
}
\value{
Numeric value of cosine similarity
}
\description{
Compute soft weighted cosine similarity of two numeric vectors
}
\details{
\code{a}, \code{b} and \code{c} are expected to have the same length.
Soft cosine similarity allows accounting for (cor)relations between features.
}
