% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{upperTMean}
\alias{upperTMean}
\title{Mean of the upper triangle of given numerical matrix}
\usage{
upperTMean(inmat, weights = NULL)
}
\arguments{
\item{inmat}{a (n x k) numerical matrix of portfolios where rows represent n subjects and columns their portfolios}
}
\description{
\code{upperTMean} computes (weighted) mean on the upper triangle of given numerical matrix.
}

