% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vCosSimilarity}
\alias{vCosSimilarity}
\title{Compute cosine similarity of two numeric vectors}
\usage{
vCosSimilarity(a, b)
}
\arguments{
\item{a}{numeric vector}

\item{b}{numeric vector}
}
\value{
Numeric value.
}
\description{
Compute cosine similarity of two numeric vectors
}
\details{
\code{a} and \code{b} are expected to have the same length.
\deqn{\sum(a*b)/\sqrt(\sum(a^2)\sum(b^2))}
}
