% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wtMCosSimilarity}
\alias{wtMCosSimilarity}
\title{Compute weighted cosine similarities for each pair of rows from given matrix and given weights}
\usage{
wtMCosSimilarity(input_mat, weights)
}
\arguments{
\item{input_mat}{numeric input matrix}

\item{weights}{numeric vector of weights}
}
\value{
Upper triangle matrix where \{i, j\} element is the cosine similarity of i-th and j-th row of the original matrix.
}
\description{
Compute weighted cosine similarities for each pair of rows from given matrix and given weights
}

