% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{aggSimilarity}
\alias{aggSimilarity}
\title{Aggregate similarity of banking porfolios.}
\usage{
aggSimilarity(porftolios, weights = NULL, rw = NULL)
}
\arguments{
\item{weights}{a (n x n) matrix of weights for weghting matrix to an average}

\item{rw}{(1 x k) vector of weights on sectors}

\item{portfolios}{a (n x k) matrix of portfolios where rows represent n subjects and columns their portfolios}
}
\value{
a numeric value of (weighted) mean similarity.
}
\description{
\code{aggSimilarity} computes aggregate similarity of banks' portfolios.
}
\details{
This is an R version of the function.
}
\seealso{
\code{\link{aggSimilarity_time}}
}

