% Generated by roxygen2 (4.0.2): do not edit by hand
\name{aggSimilarity}
\alias{aggSimilarity}
\title{Aggregate similarity of banking porfolios.}
\usage{
aggSimilarity(porftolios, weights = NULL)
}
\arguments{
\item{weights}{a (n x n) matrix of weights}

\item{portfolios}{a (n x k) matrix of portfolios where rows represent n subjects and columns their portfolios}
}
\value{
a numeric value of (weighted) mean similarity.
}
\description{
\code{aggSimilarity} computes aggregate similarity of banks' portfolios.
}
\details{
This is an R version of the function.
}
\seealso{
\code{\link{aggSimilarity_time}}
}

