% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{aggSimilarity_time}
\alias{aggSimilarity_time}
\title{Aggregate similarities of banking porfolios on a list}
\usage{
aggSimilarity_time(portfolios_list, ...)
}
\arguments{
\item{portfolios_list}{a list of portfolio matrices where rows represent n subjects and columns their portfolios}

\item{...}{other archments past to \code{\link{aggSimilarity}}}
}
\value{
a numeric vector.
}
\description{
\code{aggSimilarity_time} computes aggregate similarities for a given list of banks' portfolios
}
\details{
This is an R version of the function.
}
\seealso{
\code{\link{aggSimilarity}}
}

