% Generated by roxygen2 (4.0.2): do not edit by hand
\name{aggSimilarity_time}
\alias{aggSimilarity_time}
\title{Aggregate similarities of banking porfolios on a list}
\usage{
aggSimilarity_time(portfolios_list)
}
\arguments{
\item{portfolios_list}{a list of portfolio matrices where rows represent n subjects and columns their portfolios}
}
\value{
a numeric vector.
}
\description{
\code{aggSimilarity_time} computes aggregate similarities for a given list of banks' portfolios
}
\details{
This is an R version of the function.
}
\seealso{
\code{\link{aggSimilarity}}
}

