% Generated by roxygen2 (4.0.2): do not edit by hand
\name{upperTMean}
\alias{upperTMean}
\title{Mean of the upper triangle of given numerical matrix}
\usage{
upperTMean(inmat, weights = NULL)
}
\arguments{
\item{inmat}{a (n x k) numerical matrix of portfolios where rows represent n subjects and columns their portfolios}
}
\description{
\code{upperTMean} computes (weighted) mean on the upper triangle of given numerical matrix.
}

