% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vCosSimilarity}
\alias{vCosSimilarity}
\title{Compute cosine similarity of two numeric vectors}
\usage{
vCosSimilarity(a, b)
}
\arguments{
\item{a}{numeric vector}

\item{b}{numeric vector}
}
\value{
Numeric value of cosine similarity
}
\description{
Compute cosine similarity of two numeric vectors
}
