% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wtMCosSimilarity}
\alias{wtMCosSimilarity}
\title{Compute weighted cosine similarities for each pair of rows from given matrix and given weights}
\usage{
wtMCosSimilarity(input_mat, weights)
}
\arguments{
\item{input_mat}{numeric input matrix}

\item{weights}{numeric vector of weights}
}
\value{
Upper triangle matrix where \{i, j\} element is the cosine similarity of i-th and j-th row of the original matrix.
}
\description{
Compute weighted cosine similarities for each pair of rows from given matrix and given weights
}
\details{
\link{wtVCosSimilarity}
}
\seealso{
\code{\link{wtVCosSimilarity}}
}
